clear ram package
