+++ services/convertlog2feed/core.go
@@ -22,6 +22,7 @@ import (
 	"strings"
 	"time"
 
+	"github.com/BrunoReboul/ram/utilities/aut"
 	"github.com/BrunoReboul/ram/utilities/cai"
 	"github.com/BrunoReboul/ram/utilities/gps"
 	"github.com/BrunoReboul/ram/utilities/ram"
@@ -126,7 +127,7 @@ func Initialize(ctx context.Context, global *Global) {
 		instanceDeployment.Core.ServiceName,
 		instanceDeployment.Core.SolutionSettings.Hosting.ProjectID)
 
-	if clientOption, ok = ram.GetClientOptionAndCleanKeys(ctx,
+	if clientOption, ok = aut.GetClientOptionAndCleanKeys(ctx,
 		serviceAccountEmail,
 		keyJSONFilePath,
 		instanceDeployment.Core.SolutionSettings.Hosting.ProjectID,
