+++ services/getgroupsettings/core.go
@@ -20,6 +20,7 @@ import (
 	"fmt"
 	"log"
 
+	"github.com/BrunoReboul/ram/utilities/aut"
 	"github.com/BrunoReboul/ram/utilities/cai"
 	"github.com/BrunoReboul/ram/utilities/gps"
 	"github.com/BrunoReboul/ram/utilities/ram"
@@ -69,7 +70,7 @@ func Initialize(ctx context.Context, global *Global) {
 		instanceDeployment.Core.ServiceName,
 		instanceDeployment.Core.SolutionSettings.Hosting.ProjectID)
 
-	if clientOption, ok = ram.GetClientOptionAndCleanKeys(ctx,
+	if clientOption, ok = aut.GetClientOptionAndCleanKeys(ctx,
 		serviceAccountEmail,
 		keyJSONFilePath,
 		instanceDeployment.Core.SolutionSettings.Hosting.ProjectID,
