+++ utilities/aut/type_keyrestapiformat.go
@@ -1,0 +1,25 @@
+// Copyright 2020 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the 'License');
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an 'AS IS' BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package aut
+
+// keyRestAPIFormat Service account json key using REST API format
+type keyRestAPIFormat struct {
+	Name            string `json:"name"`
+	PrivateKeyType  string `json:"privateKeyType"`
+	PrivateKeyData  string `json:"privateKeyData"`
+	ValidAfterTime  string `json:"validAfterTime"`
+	ValidBeforeTime string `json:"validBeforeTime"`
+	KeyAlgorithm    string `json:"keyAlgorithm"`
+}
